api_tests/requirements.txt
pytest>=7.4.0
pytest-html>=3.2.0
pytest-xdist>=3.3.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
